# api/openapi/order.yaml
openapi: 3.1.0
info:
  title: order service
  description: order service
  version: 1.0.0
servers:
  - url: 'https://{hostname}/api'
    variables:
      hostname:
        default: '127.0.0.1'
paths:
  /customer/{customerID}/orders/{orderID}:
    get:
      description: "get orders"
      parameters:
        - name: customerID
          in: path
          required: true
          schema:
            type: string
        - name: orderID
          in: path
          required: true
          schema:
            type: string

      responses:
        "200":
          description: todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        default:
          description: todo
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Error'

  /customer/{customerID}/orders:
    post:
      description: "create orders"
      parameters:
        - name: customerID
          in: path
          required: true
          schema:
            type: string

      requestBody:
        required: true
        content:
          application/json:
            schema:
               $ref: '#/components/schemas/CreateOrderRequest'

      responses:
        "200":
          description: todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        default:
          description: todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Order:
        type: object
        properties:
          id:
            type: string
          customerID:
            type: string
          status:
            type: string
          items:
            type: array
            items:
              $ref: '#/components/schemas/Item'
          paymentLink:
            type: string

    Item:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        quantity:
          type: integer
          format: int32
        priceID:
          type: string

    CreateOrderRequest:
      type: object
      required:
        - customerID
        - items
      properties:
        customerID:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemWithQuantity'

    ItemWithQuantity:
       type: object
       properties:
          id:
            type: string
          quantity:
            type: integer
            format: int32

    Error:
      type: object
      properties:
        message:
            type: string